+++ Assets/AppCenter/AppCenterBehavior.cs
@@ -7,6 +7,7 @@ using System;
 using System.Reflection;
 using Microsoft.AppCenter.Unity.Internal;
 using System.Linq;
+using Microsoft.AppCenter.Unity.Analytics;
 
 [HelpURL("https://docs.microsoft.com/en-us/appcenter/sdk/crashes/unity")]
 public class AppCenterBehavior : MonoBehaviour
@@ -75,6 +76,9 @@ public class AppCenterBehavior : MonoBehaviour
         if (IsStartFromAppCenterBehavior(advancedSettings))
         {
             AppCenter.LogLevel = Settings.InitialLogLevel;
+            if (Settings.EnableManualSessionTracker) {
+                Analytics.EnableManualSessionTracker();
+            }
             if (Settings.CustomLogUrl.UseCustomUrl)
             {
                 AppCenter.SetLogUrl(Settings.CustomLogUrl.Url);