+++ Assets/AppCenter/Editor/AppCenterPreBuild.cs
@@ -115,6 +115,7 @@ public class AppCenterPreBuild : IPreprocessBuildWithReport
         settingsMaker.SetAppSecret(settings);
         settingsMaker.SetLogLevel((int)settings.InitialLogLevel);
         settingsMaker.IsAllowNetworkRequests((bool)settings.AllowNetworkRequests);
+        settingsMaker.EnableManualSessionTracker((bool)settings.EnableManualSessionTracker);
         if (settings.CustomLogUrl.UseCustomUrl)
         {
             settingsMaker.SetLogUrl(settings.CustomLogUrl.Url);