+++ Assets/AppCenter/Editor/IAppCenterSettingsMaker.cs
@@ -23,4 +23,5 @@ public interface IAppCenterSettingsMaker
     void SetShouldEnableDistributeForDebuggableBuild();
     void SetDistributeDisableAutomaticCheckForUpdate();
     void SetUpdateTrack(int updateTrack);
+    void EnableManualSessionTracker(bool isEnabled);
 }